Update TC.py
Adding NatNet UDP to have the test controller also communicate with the OptiTrack Motion Capture system. The goal is to have the gantry move with predetermined coordinates and have the OptiTrack measure the positional data during the movement. The data is streamed to the test controller, where it will be compared with the positional data that is being read from the Open Builds GUI. This step is mainly for establishing a "ground truth", where we have an accurate account of the devices position in an environment which is crucial when studying signal characteristics of dynamic devices.